Implemented UI and E2E completion

+++ src/main/java/com/revival/weatherApp/service/WeatherService.java
@@ -3,9 +3,10 @@ package com.revival.weatherApp.service;
 import com.revival.weatherApp.model.WeatherModel;
 
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.util.Map;
 
 public interface WeatherService {
     WeatherModel getWeatherData(String city, String unit) throws IOException;
-    Map<String, Double> getGeoCoordinates(String city);
+    Map<String, BigDecimal> getGeoCoordinates(String city);
 }
